feat : 클래스룸에 과정 등록 구현
