+++ Backend/edututor/src/main/java/com/tls/edututor/course/course/entity/CourseClassroom.java
@@ -1,13 +1,16 @@
 package com.tls.edututor.course.course.entity;
 
 import com.tls.edututor.classroom.entity.Classroom;
+import com.tls.edututor.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
+@Setter
 @Entity
 @Table(name = "COURSE_CLASSROOM")
-public class CourseClassroom {
+public class CourseClassroom extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
