+++ Backend/edututor/src/main/java/com/tls/edututor/course/courseclassroom/repository/CourseClassroomRepository.java
@@ -4,4 +4,6 @@ import com.tls.edututor.course.course.entity.CourseClassroom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CourseClassroomRepository extends JpaRepository<CourseClassroom, Long> {
+  boolean existsByCourseIdAndClassroomId(Long courseId, Long classroomId);
+
 }
