--- Frontend/edututor/src/assets/css/CourseFilterPage.css
@@ -1,40 +1,0 @@
-.filter-options {
-    display: flex;
-    gap: 1rem;
-    margin-bottom: 1rem;
-}
-
-.filter-group {
-    display: flex;
-    flex-direction: column;
-}
-
-button {
-    padding: 0.5rem;
-    border: none;
-    background-color: #f0f0f0;
-    cursor: pointer;
-    margin: 0.25rem 0;
-}
-
-button.selected {
-    background-color: #007bff;
-    color: #fff;
-}
-
-.filtered-courses {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.course-list {
-    display: flex;
-    gap: 1rem;
-    flex-wrap: wrap;
-}
-
-.course-item {
-    width: 150px;
-    text-align: center;
-}
